pj/clothing_segmentation/requirements.txt
torch
torchvision
gradio
gdown==4.6.0
Pillow
opencv-python
numpy
